--- 3380 Project/src/components/CreationForm.tsx
@@ -1,1 +1,0 @@
-
