--- 3380 Project/src/contexts/CategoryContext.tsx
@@ -1,31 +1,0 @@
-import React, { createContext, useState, useContext, ReactNode } from 'react';
-
-interface Category {
-  id: number;
-  element: JSX.Element;
-}
-
-interface CategoryContextType {
-  categories: Category[];
-  setCategories: React.Dispatch<React.SetStateAction<Category[]>>;
-}
-
-const CategoryContext = createContext<CategoryContextType | undefined>(undefined);
-
-export const CategoryProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
-  const [categories, setCategories] = useState<Category[]>([]);
-
-  return (
-    <CategoryContext.Provider value={{ categories, setCategories }}>
-      {children}
-    </CategoryContext.Provider>
-  );
-};
-
-export const useCategories = () => {
-  const context = useContext(CategoryContext);
-  if (context === undefined) {
-    throw new Error('useCategories must be used within a CategoryProvider');
-  }
-  return context;
-};
